sysoptions.py: p7zip: allow users to disable integrity check
p7zip integrity check will basically double the decompression time, this is
not a big problem on modern systems with SSD/NVME and a performant CPU, but
can be a pain on old systems with HDD. For this reason better add an option
and allow user with low-end devices to disable p7zip verify.

Note: with integrity disabled an exception will still be raised if the
decompression fail, but p7zip will decompress the files even if corrupted.
In this case Lutris may recognize the runner as properly installed even if
there are corrupted files exctracted, that's why I suggest to keep it
enabled unless necessary.

Signed-off-by: Marco Zanin <mrczn.bb@gmail.com>