[1/2]xrandr: cache available resolutions and outputs
Compared to other tabs, system is really slow the first time user open
it in game/runner settings and preferences. The main cause is xrandr.
It runs everytime and the window wait for xrandr output, until is ready
system settings box won't show up. Cache the result greatly helps in
this situation.

Test: manually tested available resolutions.

Signed-off-by: Marco Zanin <mrczn.bb@gmail.com>