[4/4]Configs: add a class method for reset istances
When config is saved, especially for a low level like system the
other configs must be updated accordingly. This commit adds a method
to reset istances when needed. A change in system level will reset all
istances, a change on runners will reset only games istances while a
change on games does not reset anything

Test: saved on system and runner levels, needed istances are cleared
and changes propagated

Signed-off-by: Marco Zanin <mrczn.bb@gmail.com>