Sysoptions: do not show disable_compositor in Plasma sessions
Commit 80ff3b06ad8cc3c3a1db6701ca90e8f09decbb3e
"Remove qdbus usage / Plasma users know better than stopping their
compositors and have better computers than i3 and XFCE users." Did not
take in consideration the fact that now on Plasma with X11 the option
is still visible, but it's greyed out and there's no way to enable it.
This commit adds (and cache) a simple check for Plasma desktop and if
True the option is not visible.

Test: tested on Plasma desktop with X11, option is not visible

Signed-off-by: Marco Zanin <mrczn.bb@gmail.com>